first draft of script to extract source info
Barring the use of an actual database, we would like to be able to
access all the sources and their associated verses easily.  This is a
simple state machine parser that allows us to extract all the sources.
It needs a little refinement, but the basic skeleton is now in place.