output yaml that requires no dependencies
The best way to work with data like this from within ruby is probably
using ruby structs (dot access for attributes and highly memory
efficient).  However, if we emit yaml on ruby structs we get yaml output
that cannot be read in with ruby unless the structs are already defined.
This isn't a problem if you know what you are doing (i.e., simply
require the extract_sources.rb file first) but makes everything less
user friendly.  Converting to hashes from structs results in symbol keys
which are fine with ruby but awkard to use from other languages (e.g.,
python).  So, I default the output to string-keyed hashes which can be
read in by ruby (without dependencies) and will play nicely with other
languages.  I've left the option to output ruby structs in the yaml for
those that prefer this.

Also, changed name of Verse to ScriptureVerse to avoid ambiguity with the
verse attribute.